Wayland: Area margins

+++ src/wayland/regionMargins/QvkRegionMargins_wl.cpp
@@ -154,7 +154,7 @@ void QvkRegionMargins_wl::slot_snapshot()
             bus.connect( "", reply.value().path(), "org.freedesktop.portal.Request", "Response", this, SLOT( slot_handle_response_snapshot(uint,QVariantMap) ) );
             qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] slot_snapshot()" << reply.value().path();
         } else {
-            qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] slot_pushButton_snapshot() Something is wrong: " << reply.error();
+            qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] slot_snapshot() Something is wrong: " << reply.error();
         }
 }
 
@@ -166,60 +166,65 @@ void QvkRegionMargins_wl::slot_handle_response_snapshot( uint responseCode, QVar
         QFileInfo fileInfo( url.toLocalFile() );
         QString filePath_org = fileInfo.absoluteFilePath();
 
-        /*
+        if ( filePath_org > "" ) {
+
+            /*
          * Hinweis für alle Ränder
          * Die Pixel werden von der Mitte zum Rand hin ausgwertet.
          *
          */
-
-        QImage image = QImage( filePath_org );
-        for( int i = image.height()/2; i > 0; i-- ) {
-            QColor pixelColor = image.pixelColor( image.width()/2, i );
-            if ( pixelColor == color ) {
-                top = i + 1 - lineWidth;
-                qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] Top margin:" << top;
-                break;
+            
+            QImage image = QImage( filePath_org );
+            for( int i = image.height()/2; i > 0; i-- ) {
+                QColor pixelColor = image.pixelColor( image.width()/2, i );
+                if ( pixelColor == color ) {
+                    top = i + 1 - lineWidth;
+                    qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] Top margin:" << top;
+                    break;
+                }
             }
-        }
-
-        for( int i = image.width()/2; i < image.width(); i++ ) {
-             QColor pixelColor = image.pixelColor( i, image.height()/2 );
-             if ( pixelColor == color ) {
-                 right = image.width() - i - lineWidth;
-                 qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] right margin:" << right;
-                 break;
-             }
-         }
-
-        for( int i = image.height()/2; i < image.height(); i++ ) {
-             QColor pixelColor = image.pixelColor( image.width()/2, i );
-             if ( pixelColor == color ) {
-                 bottom = image.height() - i - lineWidth;
-                 qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] bottom margin:" << bottom;
-                 break;
-             }
-         }
-
-        for( int i = image.width()/2; i > 0; i-- ) {
-            QColor pixelColor = image.pixelColor( i, image.height()/2 );
-            if ( pixelColor == color ) {
-                left = i + 1 - lineWidth;
-                qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] Left margin:" << left;
-                break;
+            
+            for( int i = image.width()/2; i < image.width(); i++ ) {
+                QColor pixelColor = image.pixelColor( i, image.height()/2 );
+                if ( pixelColor == color ) {
+                    right = image.width() - i - lineWidth;
+                    qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] right margin:" << right;
+                    break;
+                }
             }
+            
+            for( int i = image.height()/2; i < image.height(); i++ ) {
+                QColor pixelColor = image.pixelColor( image.width()/2, i );
+                if ( pixelColor == color ) {
+                    bottom = image.height() - i - lineWidth;
+                    qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] bottom margin:" << bottom;
+                    break;
+                }
+            }
+            
+            for( int i = image.width()/2; i > 0; i-- ) {
+                QColor pixelColor = image.pixelColor( i, image.height()/2 );
+                if ( pixelColor == color ) {
+                    left = i + 1 - lineWidth;
+                    qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] Left margin:" << left;
+                    break;
+                }
+            }
+            
+            QFile file( filePath_org );
+            file.remove();
+
+            qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] slot_handle_response_snapshot()" << size();
+
+            setVisible( false );
+
+            emit signal_regionMargins();
         }
-
-        QFile file( filePath_org );
-        file.remove();
-
-        qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] slot_handle_response_snapshot()" << size();
     } else {
+        setVisible( false );
         qDebug().noquote() << global::nameOutput << "[QvkRegionMargins_wl] slot_handle_response_snapshot() Unable to take a screenshot" << results["uri"];
     }
 
-    setVisible( false );
-
-    emit signal_regionMargins();
 }
